refactor: 백엔드 코드 구조 개선 및 기능별 파일 분리 (config, schemas, services)
